# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Beyond HOA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /violations:
    get:
      operationId: listViolations
      tags: [violations]
      summary: List violations
      responses:
        "200":
          description: List of violations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Violation"
    post:
      operationId: createViolation
      tags: [violations]
      summary: Create a violation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViolationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"
  /violations/{id}/status:
    put:
      operationId: updateViolationStatus
      tags: [violations]
      summary: Update a violation's status
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViolationStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"
  /violations/{id}:
    delete:
      operationId: deleteViolation
      tags: [violations]
      summary: Delete a violation
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /violations/analyze-image:
    post:
      operationId: analyzeViolationImage
      tags: [violations]
      summary: Analyze a violation photo with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViolationImageInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViolationAnalysis"
...
  /work-orders:
    get:
      operationId: listWorkOrders
      tags: [work-orders]
      summary: List work orders
      responses:
        "200":
          description: List of work orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
    post:
      operationId: createWorkOrder
      tags: [work-orders]
      summary: Create a work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /work-orders/{id}:
    put:
      operationId: updateWorkOrder
      tags: [work-orders]
      summary: Update a work order
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
    delete:
      operationId: deleteWorkOrder
      tags: [work-orders]
      summary: Delete a work order
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Violation:
      type: object
      required:
        - id
        - resident_name
        - unit
        - violation_type
        - notice_number
        - incident_date
        - description
        - required_action
        - compliance_deadline
        - status
        - created_at
      properties:
        id: { type: integer }
        resident_name: { type: string }
        unit: { type: string }
        violation_type: { type: string }
        notice_number: { type: integer }
        incident_date: { type: string }
        description: { type: string }
        required_action: { type: string }
        compliance_deadline: { type: string }
        fine_amount: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        issued_by: { type: ["string", "null"] }
        photo_url: { type: ["string", "null"] }
        assigned_vendor: { type: ["string", "null"] }
        status: { type: string, enum: [open, resolved, appealed] }
        created_at: { type: string, format: date-time }
    ViolationInput:
      type: object
      required:
        - resident_name
        - unit
        - violation_type
        - incident_date
        - description
        - required_action
        - compliance_deadline
      properties:
        resident_name: { type: string, minLength: 1 }
        unit: { type: string, minLength: 1 }
        violation_type: { type: string, minLength: 1 }
        notice_number: { type: integer }
        incident_date: { type: string }
        description: { type: string, minLength: 1 }
        required_action: { type: string, minLength: 1 }
        compliance_deadline: { type: string }
        fine_amount: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        issued_by: { type: ["string", "null"] }
        photo_url: { type: ["string", "null"] }
        assigned_vendor: { type: ["string", "null"] }
    ViolationStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [open, resolved, appealed] }
    ViolationImageInput:
      type: object
      required: [imageBase64]
      properties:
        imageBase64: { type: string }
        mimeType: { type: string }
    ViolationAnalysis:
      type: object
      required:
        - violation_type
        - description
        - required_action
        - severity
        - compliance_days
        - summary
      properties:
        violation_type: { type: string }
        description: { type: string }
        required_action: { type: string }
        severity: { type: string, enum: [low, medium, high] }
        fine_suggestion: { type: ["number", "null"] }
        compliance_days: { type: integer }
        summary: { type: string }
...
    WorkOrder:
      type: object
      required:
        - id
        - title
        - resident_name
        - unit
        - category
        - priority
        - description
        - status
        - created_at
        - updated_at
      properties:
        id: { type: integer }
        title: { type: string }
        resident_name: { type: string }
        unit: { type: string }
        category: { type: string }
        priority: { type: string }
        description: { type: string }
        status: { type: string, enum: [submitted, "in-progress", completed, cancelled] }
        board_notes: { type: ["string", "null"] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WorkOrderInput:
      type: object
      required: [title, resident_name, unit, category, description]
      properties:
        title: { type: string, minLength: 1 }
        resident_name: { type: string, minLength: 1 }
        unit: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        priority: { type: string }
        description: { type: string, minLength: 1 }
    WorkOrderUpdate:
      type: object
      properties:
        status: { type: string, enum: [submitted, "in-progress", completed, cancelled] }
        board_notes: { type: ["string", "null"] }
        priority: { type: string }
...