# lib/api-zod/src/generated/types/violationInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Beyond HOA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ViolationInput {
  /** @minLength 1 */
  resident_name: string;
  /** @minLength 1 */
  unit: string;
  /** @minLength 1 */
  violation_type: string;
  notice_number?: number;
  incident_date: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  required_action: string;
  compliance_deadline: string;
  /** @nullable */
  fine_amount?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  issued_by?: string | null;
  /** @nullable */
  photo_url?: string | null;
  /** @nullable */
  assigned_vendor?: string | null;
}